Split BackendBinding from example/Utils into src/utils
This will make it possible to reuse the backend bindings for test suites
