+++ examples/CMakeLists.txt
@@ -12,67 +12,52 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-list(APPEND UTILS_SOURCES
-    Utils.h
+add_library(sample_utils
     Utils.cpp
-    BackendBinding.h
+    Utils.h
 )
-
-if (APPLE)
-    list(APPEND UTILS_SOURCES
-        MetalBinding.mm
-    )
-endif()
-
-if (WIN32)
-    list(APPEND UTILS_SOURCES
-        D3D12Binding.cpp
-    )
-endif()
-
-add_library(utils STATIC ${UTILS_SOURCES})
-target_link_libraries(utils nxt_backend nxt_wire shaderc nxtcpp nxt)
-SetCXX14(utils)
+target_link_libraries(sample_utils utils nxt_wire)
+SetCXX14(sample_utils)
 
 add_executable(CHelloTriangle CHelloTriangle.cpp)
-target_link_libraries(CHelloTriangle utils)
+target_link_libraries(CHelloTriangle sample_utils)
 SetCXX14(CHelloTriangle)
 
 add_executable(CppHelloTriangle HelloTriangle.cpp)
-target_link_libraries(CppHelloTriangle utils)
+target_link_libraries(CppHelloTriangle sample_utils)
 SetCXX14(CppHelloTriangle)
 
 add_executable(ComputeBoids ComputeBoids.cpp)
-target_link_libraries(ComputeBoids utils)
+target_link_libraries(ComputeBoids sample_utils)
 target_include_directories(ComputeBoids PUBLIC ../ ${GLM_INCLUDE_DIR})
 SetCXX14(ComputeBoids)
 
 add_executable(HelloVertices HelloVertices.cpp)
-target_link_libraries(HelloVertices utils)
+target_link_libraries(HelloVertices sample_utils)
 SetCXX14(HelloVertices)
 
 add_executable(HelloInstancing HelloInstancing.cpp)
-target_link_libraries(HelloInstancing utils)
+target_link_libraries(HelloInstancing sample_utils)
 SetCXX14(HelloInstancing)
 
 add_executable(HelloIndices HelloIndices.cpp)
-target_link_libraries(HelloIndices utils)
+target_link_libraries(HelloIndices sample_utils)
 SetCXX14(HelloIndices)
 
 add_executable(HelloUBO HelloUBO.cpp)
-target_link_libraries(HelloUBO utils)
+target_link_libraries(HelloUBO sample_utils)
 SetCXX14(HelloUBO)
 
 add_executable(HelloCompute HelloCompute.cpp)
-target_link_libraries(HelloCompute utils)
+target_link_libraries(HelloCompute sample_utils)
 SetCXX14(HelloCompute)
 
 add_executable(RenderToTexture RenderToTexture.cpp)
-target_link_libraries(RenderToTexture utils)
+target_link_libraries(RenderToTexture sample_utils)
 SetCXX14(RenderToTexture)
 
 add_executable(Animometer Animometer.cpp)
-target_link_libraries(Animometer utils)
+target_link_libraries(Animometer sample_utils)
 SetCXX14(Animometer)
 
 add_executable(SpirvTest SpirvTest.cpp)
@@ -80,7 +65,7 @@ target_link_libraries(SpirvTest shaderc spirv-cross nxtcpp)
 SetCXX14(SpirvTest)
 
 add_executable(CppHelloDepthStencil HelloDepthStencil.cpp)
-target_link_libraries(CppHelloDepthStencil utils)
+target_link_libraries(CppHelloDepthStencil sample_utils)
 SetCXX14(CppHelloDepthStencil)
 
 add_subdirectory(glTFViewer)
